3D: Added Index/Staging Buffers
Models can now use index buffers as well and the program will automatically check for them. We also added staging buffers to our create buffers process so that we can use local memory from the GPU to our main buffer that is also in the GPU. This will give us the best performance in the long run, but is best used for static objects. Also tweaked our camera controls