3D: Added Point Lights and Attenuation
Set up a point light in the GlobalUbo. We use the inverse square law in the vertex shader to shade the vertices based on the distance of the light, as well as the angle at which the light hits it. Also added a plane for the rendered objects to sit on. Also changed the view distance for the camera to be 100 units.... True story