Fragment Lighting
Added a feature to the vertex shader where it interpolates between the vertex normals. I use this to determine the normals of each fragment between the vertices and be able to apply lighting based on angle and intensity. The blotchy shadows are now gone, and objects look as intended even if their vertices are far apart. The fragment shader colors each pixel accordingly.